Split TrustedRouter safety client
