+++ docs/DECISIONS.md
@@ -5,6 +5,7 @@
 - Codex-style workspace chrome should keep model selection and approval mode as separate controls near the composer, not in the top bar. Model is a long-lived preference, while approval mode is an autonomy/safety posture users may change at send time. The model picker should only select models; Auto/Review/Read-only lives in a compact mode control beside it. Instruction files, memories, Computer Use readiness, and runtime issues remain accessible in the status popover and transcript/settings surfaces, but they should not occupy permanent top-bar width. This keeps the first read calmer while preserving diagnostics and Playwright coverage.
 - TrustedRouter prompt/message construction lives in `TrustedRouterPromptBuilder`, not the network client. The builder owns system prompt copy, tool schema formatting, project instruction context, memory context, message-history projection, tool-output projection, and the history limit. `TrustedRouterLLMClient` owns SDK calls and stream collection.
 - TrustedRouter API-key resolution lives in `TrustedRouterAPIKeyResolver`. Developer override precedence, whitespace trimming, session-store fallback, and the actionable missing-key error should stay there so action and safety clients cannot drift.
+- TrustedRouter action and safety transports are separate files. `TrustedRouterLLMClient` owns action streaming; `TrustedRouterSafetyModelClient` owns Auto-review model calls. They share API-key resolution through `TrustedRouterAPIKeyResolver` and JSON response parameters only while both endpoints require JSON-object responses.
 - Tool-router feature families should delegate before they sprawl. `ToolRouter` owns the common shell/file/patch entry point and composes git definitions from `GitToolCallDispatcher`; local git, GitHub PR, and worktree tool-call argument mapping lives in that dispatcher. This keeps the shared router small while preserving one public `ToolRouter.execute` API for the agent runtime.
 - Shell request policy is a tool-family concern. `ShellToolCallDispatcher` owns `host.shell.run` definitions, cwd containment, timeout bounds, environment override validation, and `ShellExecutionRequest` construction; `ToolRouter` only delegates shell calls before handling file and patch primitives.
